+++ gon/shaped/polygon.py
@@ -39,7 +39,9 @@ from gon.compound import (Compound,
                           Shaped)
 from gon.core.arithmetic import (non_negative_min,
                                  robust_divide)
-from gon.core.iterable import flatten
+from gon.core.iterable import (flatten,
+                               to_pairs_iterable,
+                               to_pairs_sequence)
 from gon.degenerate import EMPTY
 from gon.discrete import Multipoint
 from gon.geometry import Geometry
@@ -60,9 +62,7 @@ from gon.linear.raw import (raw_segment_point_distance,
                             raw_segments_distance,
                             squared_raw_point_segment_distance,
                             squared_raw_segments_distance)
-from gon.linear.utils import (from_raw_multisegment,
-                              to_pairs_iterable,
-                              to_pairs_sequence)
+from gon.linear.utils import from_raw_multisegment
 from gon.primitive import (Point,
                            RawPoint)
 from gon.primitive.point import point_to_step
