+++ tests/utils.py
@@ -21,12 +21,12 @@ from hypothesis.strategies import SearchStrategy
 
 from gon.compound import (Compound,
                           Relation)
+from gon.core.iterable import shift_sequence
 from gon.discrete import Multipoint
 from gon.hints import Coordinate
 from gon.linear import (Contour,
                         Multisegment,
                         Segment)
-from gon.linear.utils import shift_sequence
 from gon.mixed import Mix
 from gon.primitive import Point
 from gon.shaped import Multipolygon
